+++ dwds/lib/src/handlers/injector.dart
@@ -30,23 +30,21 @@ const _clientScript = 'dwds/src/injected/client';
 /// to include the injected DWDS client, enabling debugging capabilities
 /// and source mapping when running in a browser environment.
 ///
-/// The `_runMainAtStart` flag determines whether main() execution should be
-/// deferred or run immediately:
-/// - When `true`, main() is executed immediately when the app is loaded.
-/// - When `false`, main() execution is deferred, allowing for setup or
-///   debugging initialization before main() runs.
-///
-/// This separation allows for scenarios where debugging is not needed or
-/// should be explicitly avoided.
+/// TODO(yjessy): Remove this when the DWDS WebSocket connection is implemented.
+/// The `_useDwdsWebSocketConnection` flag determines the communication protocol:
+/// - When `true`, uses a socket-based implementation.
+/// - When `false`, uses Chrome-based communication protocol.
 class DwdsInjector {
   final Future<String>? _extensionUri;
   final _devHandlerPaths = StreamController<String>();
   final _logger = Logger('DwdsInjector');
-  final bool _runMainAtStart;
+  final bool _useDwdsWebSocketConnection;
 
-  DwdsInjector({Future<String>? extensionUri, bool runMainAtStart = true})
-    : _extensionUri = extensionUri,
-      _runMainAtStart = runMainAtStart;
+  DwdsInjector({
+    Future<String>? extensionUri,
+    bool useDwdsWebSocketConnection = false,
+  }) : _extensionUri = extensionUri,
+       _useDwdsWebSocketConnection = useDwdsWebSocketConnection;
 
   /// Returns the embedded dev handler paths.
   ///
@@ -115,7 +113,7 @@ class DwdsInjector {
             devHandlerPath,
             entrypoint,
             await _extensionUri,
-            _runMainAtStart,
+            _useDwdsWebSocketConnection,
           );
           body += await globalToolConfiguration.loadStrategy.bootstrapFor(
             entrypoint,
@@ -152,7 +150,7 @@ Future<String> _injectClientAndHoistMain(
   String devHandlerPath,
   String entrypointPath,
   String? extensionUri,
-  bool runMainAtStart,
+  bool useDwdsWebSocketConnection,
 ) async {
   final bodyLines = body.split('\n');
   final extensionIndex = bodyLines.indexWhere(
@@ -171,7 +169,7 @@ Future<String> _injectClientAndHoistMain(
     devHandlerPath,
     entrypointPath,
     extensionUri,
-    runMainAtStart,
+    useDwdsWebSocketConnection,
   );
   result += '''
   // Injected by dwds for debugging support.
@@ -203,7 +201,7 @@ Future<String> _injectedClientSnippet(
   String devHandlerPath,
   String entrypointPath,
   String? extensionUri,
-  bool runMainAtStart,
+  bool useDwdsWebSocketConnection,
 ) async {
   final loadStrategy = globalToolConfiguration.loadStrategy;
   final buildSettings = loadStrategy.buildSettings;
@@ -222,7 +220,7 @@ Future<String> _injectedClientSnippet(
       'window.\$dartEmitDebugEvents = ${debugSettings.emitDebugEvents};\n'
       'window.\$isInternalBuild = ${appMetadata.isInternalBuild};\n'
       'window.\$isFlutterApp = ${buildSettings.isFlutterApp};\n'
-      'window.\$runMainAtStart = $runMainAtStart;\n'
+      'window.\$useDwdsWebSocketConnection = $useDwdsWebSocketConnection;\n'
       '${loadStrategy is DdcLibraryBundleStrategy ? 'window.\$hotReloadSourcesPath = "${loadStrategy.hotReloadSourcesUri.toString()}";\n' : ''}'
       '${loadStrategy.loadClientSnippet(_clientScript)}';